Address AbstractVecOrMat through extra dispatch on AbstractVec
both for `frule!!` and `rrule!!`.
In particular, I have verified that since vectors are ultimately just
interpreted based on their start pointer we can just always cast to
column matrix, instead of trying to think whether we should cast to
column or row matrix.

It's still unclear whether this will also scale well to functions other
than gemv!, but I think it will.
Also, I have not yet figured out where to add the tests for the frule!!.

Allow complex eltypes for the `frule!!`

When rebasing we missed some of the transitions from `BlasRealFloat` to
`BlasFloat` as the allows eltypes. We fix this here for the new `gemv!` dispatches.